// backend/src/product/province.enum.ts
export enum ProvinceType {
  PINAR_DEL_RIO = 'Pinar del Río',
  ARTEMISA = 'Artemisa',
  LA_HABANA = 'La Habana',
  MAYABEQUE = 'Mayabeque',
  MATANZAS = 'Matanzas',
  CIENFUEGOS = 'Cienfuegos',
  VILLA_CLARA = 'Villa Clara',
  SANCTI_SPIRITUS = 'Sancti Spíritus',
  CIEGO_DE_AVILA = 'Ciego de Ávila',
  CAMAGUEY = 'Camagüey',
  LAS_TUNAS = 'Las Tunas',
  HOLGUIN = 'Holguín',
  GRANMA = 'Granma',
  SANTIAGO_DE_CUBA = 'Santiago de Cuba',
  GUANTANAMO = 'Guantánamo',
  ISLA_DE_LA_JUVENTUD = 'Isla de la Juventud',
}